requirements.txt
click==8.1.3
email-validator==1.3.1
Flask==2.2.2
Flask-Migrate==4.0.4
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
flask_mail==0.9.1
flask_jwt_extended==4.4.4
pipreqs==0.4.11
requests==2.28.2
SQLAlchemy==2.0.1
Werkzeug==2.2.2
gunicorn
pillow==10.2.0